requirements/base.txt
Django==5.1.6
django-environ
psycopg2-binary
djangorestframework==3.15.2

celery
django-celery-results
django-celery-beat

whitenoise

django-filter
django-extensions
django-cors-headers
django-storages

drf-jwt

boto3
attrs

djangorestframework-simplejwt
drf-spectacular

django-redis
Faker
factory-boy
pytest
pytest-django
